+++ src/megaplan_sdk/__init__.py
@@ -40,7 +40,9 @@ from megaplan_sdk.models.filter import (
     TradeFilter,
     UserSetting,
 )
+from megaplan_sdk.models.group import Group
 from megaplan_sdk.models.milestone import Milestone
+from megaplan_sdk.models.participant import Participant, parse_participant, parse_participants
 from megaplan_sdk.models.project import Project, ProjectFullDetails
 from megaplan_sdk.models.task import Task, TaskFullDetails
 from megaplan_sdk.resources.filters import FiltersResource
@@ -69,7 +71,11 @@ __all__ = [
     "ContractorHuman",
     "Employee",
     "Department",
+    "Group",
     "Milestone",
+    "Participant",
+    "parse_participant",
+    "parse_participants",
     "DateTime",
     # Filters
     "FiltersResource",