+++ src/megaplan_sdk/models/__init__.py
@@ -3,6 +3,8 @@
 from megaplan_sdk.models.base import BaseEntity
 from megaplan_sdk.models.common import File, Meta, Pagination, SortField
 from megaplan_sdk.models.deal import Deal, ProgramState, TradeFilter
+from megaplan_sdk.models.group import Group
+from megaplan_sdk.models.participant import Participant, parse_participant, parse_participants
 from megaplan_sdk.models.project import Project, ProjectFilter
 from megaplan_sdk.models.task import Task, TaskFilter
 
@@ -19,4 +21,8 @@ __all__ = [
     "Deal",
     "TradeFilter",
     "ProgramState",
+    "Group",
+    "Participant",
+    "parse_participant",
+    "parse_participants",
 ]